Simplify pinning again.
OK. Pinning is a thing. But there's many Python versions tested too.

And there's multiple OSes in the CI.

Happen that on Windows, matplotlib 3.1 does not build in Github CI, I
don't want to know why.

There's a wheel for matplotlib 3.1 for Python 3.6 and 3.7, so the CI
pass for the tuple (windows, Python 3.7, matplotlib 3.1), and fail for
all other Python versions.

I don't want to end up tailoring the list of dependencies by Python
version just to fit the specificities of GH CI. It would also looks
like I'm doing the work of the package manager, I don't want to do it,
let the computer do computer things.